refactor(contracts): move lane types to the shared contracts barrel
