+++ controller/src/modules/studio/lane-switch.ts
@@ -11,14 +11,18 @@ import { parseBooleanFlag } from "../../core/validation";
 import { Event, type EventManager } from "../system/event-manager";
 import type { ExclusiveLane, ResidentLane } from "../../services/lane-identity";
 import { exclusiveLaneOf } from "../../services/lane-identity";
-
-export type LaneSwitchState = "idle" | "running" | "ready" | "failed" | "restoring";
-
-export type LaneProbeView = {
-  ready: boolean;
-  port: number;
-  model_ids: string[];
-};
+import type {
+  LaneProbeView,
+  LaneStatus,
+  LaneSwitchJobView,
+  LaneSwitchState,
+} from "@local-studio/contracts/lanes";
+export type {
+  LaneProbeView,
+  LaneStatus,
+  LaneSwitchJobView,
+  LaneSwitchState,
+} from "@local-studio/contracts/lanes";
 
 export type LaneResidency = {
   resident_lane: ResidentLane;
@@ -31,29 +35,6 @@ export type ExclusiveModelRow = {
   lane: ExclusiveLane;
   active: boolean;
   max_model_len?: number;
-};
-
-export type LaneSwitchJobView = {
-  id: string | null;
-  state: LaneSwitchState;
-  from_lane: ResidentLane | null;
-  to_lane: ExclusiveLane | null;
-  script: string | null;
-  exit_code: number | null;
-  message: string | null;
-  error: string | null;
-  started_at: string | null;
-  finished_at: string | null;
-  idempotent?: boolean;
-};
-
-export type LaneStatus = {
-  enabled: boolean;
-  configured: boolean;
-  resident_lane: ResidentLane;
-  omlx: LaneProbeView;
-  ds4: LaneProbeView;
-  switch: LaneSwitchJobView;
 };
 
 export type LaneSwitchAcceptResult =
